tesisTomba/requirements.txt
Django==3.2.12
djangorestframework==3.14.0
psycopg2-binary==2.9.6
requests==2.25.1
qrcode==7.4.2
PyJWT==2.3.0
pytz==2022.1
django-extensions==3.2.3
gunicorn==23.0.0
djangorestframework-simplejwt==5.2.2
Werkzeug==2.3.5
pyOpenSSL==23.2.0
Pillow==9.0.1